eNose/enoseconfig/requirements.txt
aniso8601==9.0.1
APScheduler==3.10.1
bcrypt==4.0.1
blinker==1.6.2
click==8.1.3
colorama==0.4.6
Flask==2.3.2
Flask-APScheduler==1.12.4
Flask-Bcrypt==1.0.1
Flask-Cors==3.0.10
Flask-JWT-Extended==4.5.2
Flask-RESTful==0.3.10
Flask-Simple-Crypt==0.3.3
flask-swagger==0.2.14
gunicorn==20.1.0
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
pycryptodome==3.18.0
PyJWT==2.7.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
six==1.16.0
tzdata==2023.3
tzlocal==5.0.1
Werkzeug==2.3.4
zipp==3.15.0
paho-mqtt==2.1.0
cachetools==6.1.0
